backend/requirements.txt
# Django Core
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# CORS
django-cors-headers==4.3.0

# Filtering & Pagination
django-filter==23.5.0

# OpenAPI Documentation
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.11
dj-database-url==2.1.0

# Async Tasks & Caching
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Rate Limiting
django-ratelimit==4.1.0

# Environment Variables
python-decouple==3.8

# Production Server
gunicorn==21.2.0

# Code Quality (Development)
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

